+++ analyzer/requirements_py/dev/requirements.txt
@@ -1,5 +1,5 @@
 lxml==4.9.2
-nose==1.3.7
+pytest==7.3.1
 pycodestyle==2.7.0
 psutil==5.8.0
 portalocker==2.2.1
