+++ analyzer/tests/Makefile
@@ -6,8 +6,8 @@ PYTHON_BIN ?= python3
 
 REPO_ROOT ?= REPO_ROOT=$(ROOT)
 
-# Nose test runner configuration options.
-NOSECFG = --config .noserc
+# pytest test runner configuration options.
+PYTESTCFG = -c pytest.ini
 
 test_in_env: pycodestyle_in_env pylint_in_env test_unit_in_env test_functional_in_env test_build_logger test_tu_collector_in_env
 
@@ -33,14 +33,14 @@ pylint_in_env: venv_dev
 
 run_test:
 	$(REPO_ROOT) $(TEST_PROJECT) \
-		nosetests $(NOSECFG) ${TEST} || exit 1
+		pytest $(PYTESTCFG) ${TEST} || exit 1
 
 run_test_in_env: venv_dev
 	$(ACTIVATE_DEV_VENV) && $(REPO_ROOT) $(TEST_PROJECT) \
-		nosetests $(NOSECFG) ${TEST} || exit 1
+		pytest $(PYTESTCFG) ${TEST} || exit 1
 
-UNIT_TEST_CMD = $(REPO_ROOT) nosetests $(NOSECFG) tests/unit
-UNIT_TEST_COV_CMD = $(REPO_ROOT) coverage run -m nose $(NOSECFG) tests/unit && coverage report && coverage html
+UNIT_TEST_CMD = $(REPO_ROOT) pytest $(PYTESTCFG) tests/unit
+UNIT_TEST_COV_CMD = $(REPO_ROOT) coverage run -m pytest $(PYTESTCFG) tests/unit && coverage report && coverage html
 
 test_unit:
 	$(UNIT_TEST_CMD)
@@ -55,7 +55,7 @@ test_unit_cov_in_env: venv_dev
 	$(ACTIVATE_DEV_VENV) && $(UNIT_TEST_COV_CMD)
 
 FUNCTIONAL_TEST_CMD = $(REPO_ROOT) $(TEST_PROJECT) \
-		nosetests $(NOSECFG) tests/functional || exit 1
+		pytest $(PYTESTCFG) tests/functional || exit 1
 
 test_functional:
 	${PYTHON_BIN} $(ROOT)/scripts/test/check_clang.py || exit 1;
