+++ analyzer/tests/functional/host_check/__init__.py
@@ -7,10 +7,5 @@
 #
 # -------------------------------------------------------------------------
 
-
-def setup_package():
-    pass
-
-
-def teardown_package():
-    pass
+# This file is empty, and is only present so that this directory will form a
+# package.
