+++ analyzer/tests/functional/z3/test_z3.py
@@ -12,10 +12,12 @@
 
 from distutils import util
 import os
+import shutil
 import unittest
 import shlex
+import sys
 
-from libtest import env
+from libtest import env, project
 from libtest.codechecker import call_command
 
 NO_Z3_MESSAGE = "Z3 is not supported"
@@ -25,7 +27,71 @@ class TestSkeleton(unittest.TestCase):
 
     _ccClient = None
 
-    def setUp(self):
+    def setup_class(self):
+        """Setup the environment for testing z3."""
+
+        global TEST_WORKSPACE
+        TEST_WORKSPACE = env.get_workspace('z3')
+
+        # Set the TEST_WORKSPACE used by the tests.
+        os.environ['TEST_WORKSPACE'] = TEST_WORKSPACE
+
+        test_config = {}
+
+        test_project = 'suppress'
+
+        project_info = project.get_info(test_project)
+
+        # Copy the test project to the workspace. The tests should
+        # work only on this test project.
+        test_proj_path = os.path.join(TEST_WORKSPACE, "test_proj")
+        shutil.copytree(project.path(test_project), test_proj_path)
+
+        project_info['project_path'] = test_proj_path
+
+        test_config['test_project'] = project_info
+
+        # Suppress file should be set here if needed by the tests.
+        suppress_file = None
+
+        # Skip list file should be set here if needed by the tests.
+        skip_list_file = None
+
+        # Get an environment which should be used by the tests.
+        test_env = env.test_env(TEST_WORKSPACE)
+
+        # Create a basic CodeChecker config for the tests, this should
+        # be imported by the tests and they should only depend on these
+        # configuration options.
+        codechecker_cfg = {
+            'suppress_file': suppress_file,
+            'skip_list_file': skip_list_file,
+            'check_env': test_env,
+            'workspace': TEST_WORKSPACE,
+            'checkers': []
+        }
+
+        # Clean the test project, if needed by the tests.
+        ret = project.clean(test_project)
+        if ret:
+            sys.exit(ret)
+
+        test_config['codechecker_cfg'] = codechecker_cfg
+
+        # Export the test configuration to the workspace.
+        env.export_test_cfg(TEST_WORKSPACE, test_config)
+
+    def teardown_class(self):
+        """Clean up after the test."""
+
+        # TODO: If environment variable is set keep the workspace
+        # and print out the path.
+        global TEST_WORKSPACE
+
+        print("Removing: " + TEST_WORKSPACE)
+        shutil.rmtree(TEST_WORKSPACE, ignore_errors=True)
+
+    def setup_method(self, method):
 
         # TEST_WORKSPACE is automatically set by test package __init__.py .
         test_workspace = os.environ['TEST_WORKSPACE']
