+++ analyzer/tools/merge_clang_extdef_mappings/pytest.ini
@@ -1,0 +1,13 @@
+[pytest]
+
+addopts =
+# increase verbosity level
+  --verbose
+
+# stop running tests on first error
+# FIXME: Pretty please comment this horrible setting out, I hate it with a
+# passion
+#  --maxfail=1
+
+# do not capture stdout
+  --capture=sys
