+++ analyzer/tools/merge_clang_extdef_mappings/requirements_py/dev/requirements.txt
@@ -1,3 +1,3 @@
-nose==1.3.7
+pytest==7.3.1
 pycodestyle==2.4.0
 pylint==2.8.2
